refactor(doctor): split availability-staleness finding into its own module (module-health)
The #3274 stale-override doctor finding pushed two files over the 500-LOC
module-health cap: `cli/doctor/checks.py` (506 -> 535, already grandfathered)
and `core/availability.py` (489 -> 535, plus 11 > 10 public module functions).

Move the whole staleness feature — `override_set_at`, `stale_override_finding`,
`STALE_OVERRIDE_AGE`, and `_check_availability_override_staleness` — into a new
`cli/doctor/checks_availability.py`, co-located with its only consumer. The
availability *semantics* the finding keys on (which modes defer questions /
pause the pump) stay in core as new `Override.defers_questions` /
`Override.pauses_self_pump` predicates, mirroring the existing `Resolution`
ones, so the moved finding reads core's public surface, not the private
`_DEFERRING_MODES` / `_PAUSING_MODES` constants.

Result: availability.py 497 LOC / 9 public funcs; checks.py back to its 506
baseline; the new module 93 LOC / 2 public funcs — all under cap. The moved
pure-logic tests follow the code to `tests/teatree_cli/doctor/`, and a direct
`Override` predicate contract test lands in the core mirror. Behavior unchanged.
